remove prefix from test
